requirements.txt
python-docx
pyinstaller
pillow